+++ data/exchangerate/src/main/java/com/example/rate/model/ExchangeRateModel.kt
@@ -1,8 +1,6 @@
 package com.example.rate.model
 
-import com.example.common.model.ExchangeDetailRate
 import com.example.common.model.ExchangeRate
-import com.google.gson.annotations.SerializedName
 import java.math.BigDecimal
 
 /**
@@ -24,28 +22,4 @@ fun ExchangeRateModel.toExternalModel() = ExchangeRate(
     currencySymbol = currencySymbol,
     type = type,
     rateUsd = BigDecimal(rateUsd)
-)
-
-fun ExchangeRateDetailModel.toExternalModel() = ExchangeDetailRate(
-    id = rateDetail.id,
-    symbol = rateDetail.symbol,
-    currencySymbol = rateDetail.currencySymbol,
-    type = rateDetail.type,
-    rateUsd = BigDecimal(rateDetail.rateUsd),
-    timestamp = timestamp.toLong()
-)
-
-data class ExchangeRatesModel(
-    @SerializedName("data")
-    val rates: List<ExchangeRateModel>
-)
-
-fun ExchangeRatesModel.toExternalModel(): List<ExchangeRate> =
-    this.rates.map { it.toExternalModel() }
-
-
-data class ExchangeRateDetailModel(
-    @SerializedName("data")
-    val rateDetail: ExchangeRateModel,
-    val timestamp: String
 )
